ingestion/requirements.txt
grpcio
grpcio-tools
kafka-python
protobuf
requests